Enable switching between images via navigation component (#13)
* Enable switching between images via navigation component

In addition `index.tsx` file is now responsible for redirecting
the user to the last episode when they try to display the episode
which does not exist. Tests for `Navigation` component were added.
They check some edge cases. The update of  storybook's story for `Home` component was done.
Story was also addes for `Navigation`. Home` component and its type was renamed.
Update `netflity.toml` was done so that netlify has the access to episodes

Co-authored-by: Piotr Niełacny <piotr.nielacny@gmail.com>