Use native `Head` from nextjs
There is a problem with `<head>` that in nextjs, because content for
this tag is collected within all components and after that it is
rendered inside `<head>`. We need wrap component in `HeadManagerContext`
and wait for nextjs to execute. That is why we need to use `waitFor` for
this test.