Introduce static generation of episode page
`getServerSideProps` method was replaced with `getStaticProps`.
That method remains in `index.ts` and the behaviour for non-existing
episodes is changed. If an user tries to access the episode
which does not exist, the last episode in rendered instead of triggering
a redirection to the `home page`. `getEpisodes` method was extended by
adding episode slug as returned value so that it is easy to specify
paths that should be pre-rendered.