Introduce static generation of episode page (#23)
* Introduce static generation of episode page

`getServerSideProps` method was replaced with `getStaticProps`.
That method remains in `index.ts` and the behavior for non-existing
episodes is changed. If a user tries to access an episode
that does not exist, the last episode is rendered instead of triggering
a redirection to the `home page`. `getEpisodes` method was extended by
adding episode slug as a returned value so that it is easy to specify
paths that should be pre-rendered.