Simplify tests for the tags
The tests for `<title>` and `<meta>` were simplified in order to get rid of the warning `next-head-count` thrown by nextjs and thanks to that the simple function are tested. Those function returns title and tag meta description of given episode. As a result of that simplification the tests are separated from nextjs.